Resolve rubocop issues
Disable Style/SignalException because self.fail is a Puppet method mixed
into the provider, and isn't the same as ruby's builtin Object#fail.

Disable Layout/IndentHeredoc since we don't want to depend on external
library for formatting.

Disable Metrics/BlockNesting for the single method where it was used
(and reenable it afterwards).